完成发起项目部分

+++ atcrowdfunding-member-entity/src/main/java/com/yjn/crowd/po/ProjectPO.java
@@ -1,7 +1,8 @@
 package com.yjn.crowd.po;
 
-import java.io.Serializable;
 import lombok.Data;
+
+import java.io.Serializable;
 
 /**
  * t_project
@@ -24,7 +25,7 @@ public class ProjectPO implements Serializable {
     /**
      * 筹集金额
      */
-    private Long money;
+    private Integer money;
 
     /**
      * 筹集天数
